+++ src/Sql/Sql/Database Backup/Model/AzureSqlDatabaseBackupShortTermRetentionPolicyModel.cs
@@ -12,6 +12,8 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using System;
+
 namespace Microsoft.Azure.Commands.Sql.Backup.Model
 {
     public class AzureSqlDatabaseBackupShortTermRetentionPolicyModel
@@ -37,6 +39,11 @@ namespace Microsoft.Azure.Commands.Sql.Backup.Model
         public int RetentionDays { get; set; }
 
         /// <summary>
+        /// Gets or sets the differential backup interval of this policy.
+        /// </summary>
+        public int? DiffBackupIntervalInHours { get; set; }
+
+        /// <summary>
         /// Construct AzureSqlDatabaseBackupShortTermRetentionPolicyModel from Management.Sql.BackupShortTermRetentionPolicy object
         /// </summary>
         /// <param name="resourceGroup"></param>
@@ -44,10 +51,16 @@ namespace Microsoft.Azure.Commands.Sql.Backup.Model
         /// <param name="policy"></param>
         public AzureSqlDatabaseBackupShortTermRetentionPolicyModel(string resourceGroup, string serverName, string databaseName, Management.Sql.Models.BackupShortTermRetentionPolicy policy)
         {
+            if (policy.RetentionDays == null && policy.DiffBackupIntervalInHours == null)
+            {
+                throw new ArgumentException(string.Format(Microsoft.Azure.Commands.Sql.Properties.Resources.SetAzSqlDatabaseBackupShortTermRetentionInvalidParameters, "RetentionDays", "DiffBackupIntervalInHours"));
+            }
+
             ResourceGroupName = resourceGroup;
             ServerName = serverName;
             DatabaseName = databaseName;
             RetentionDays = policy.RetentionDays.Value;
+            DiffBackupIntervalInHours = policy.DiffBackupIntervalInHours.Value;
         }
     }
 }